Refactor library for handling trace entries.
Break out the parsing and writing of entries to a fd/string into a library.
Use this new library everywhere that messages are parsed and written out.

Refactor the blueprint file to create libmemory_replay so that the tools
don't need to compile pieces of the memory replay code.

Add new unit tests for the new library.

Remove the AndroidTest.xml since it doesn't apply any more and the host
tests are no different then the tests run on device. Also, change to
the test suites to general-tests to reflect this.

Test: All unit tests pass.
Test: Run the print_trace on zipped trace files and verify the same output as
Test: unzipped traces.
Change-Id: I389cd1445cdcd9c24038f91b5f00f66d07b8cb6c