Fix memory corruption in make resident function.
The MakeAllocationResident function was writing past the end of the
allocation. The fix is to remove the change of the data pointer.
The code is already using an index which will cause the pointer
to be page aligned.

In addition, make the code runnable on host and make the host
executable unstripped.

Test: Built the host sanitizer and ran memory_replay{32,64} on
Test: traces without any detected memory corruption.
Change-Id: Icc988bdb38ec82259271918d197782e79eefaccf