linker: Fix is-proc-mounted check to check for mountedness
Previously the check was for existence of the "/proc" path, which always
exists since it is the path over which the "proc" filesystem is mounted.

Since the purpose of the check is to test whether realpath() should
work, test for "/proc/self/fd", the directory used by real path() to
query the kernel for the "real" path, which won't exist until "proc" has
actually been mounted.

Bug: 343831934
Change-Id: Ia9ed0db2cb9688d7a47586712513119293a8a53f