arm64 ifuncs: fix oryon detection.
qcom now says that it's implementer and part that should be
checked, not implementer and variant. The previous use of
variant was a bug, which they saw because old pre-Oryon cpus
started to use the supposedly Oryon-specific routine. Variant
is actually just used to distinguish big/little cores for a
given part.

Additionally, though there's no documentation to link to,
all parts 0..15 should be assumed to be Oryon.

Add trivial "getters" for the fields of midr_el1 so --
whatever test we eventually end up with -- it's at least
as readable as possible.

Change-Id: I87d1f2e6a238c77263df292ff321d02674ff87a8