bionic: loader: 16k: Refactor Setup16KiBAppCompat()
Remove the error param from Setup16KiBAppCompat() and make it return
void instead of a boolean.

This is incorrect on it's own as the ELF layouts not eligible for
RX|RW compat will now attempt loading and crash instead of reporting an
error.

This is fixed by subsequent patches that introduce compat-loading of
ELF layouts that cannot be loaded in the RX|RW compat mode.

Bug: 418049947
Test: manual: 3000 top apps
Change-Id: I01e80b01d562a2a21273bc09a6a51a14ed378a5a
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>