Remove dead code.
Static binaries were the only user of the "generic" routines for many years, and when I fixed things so that static binaries used ifuncs too, these all became obsolete.

The OpenBSD strcat.c is actually still used for arm64, believe it or not, so although we can remove the OpenBSD stpcpy.c as no longer used, strcat.c lives on!

Change-Id: I2b7e4bfaa2c6060511fbec52fcafe2eb9c4d5590