linker: Use load_bias in ElfReader::FindGnuPropertySection
Previously load_start was used instead, which may not equal load_bias
if the first load segment's virtual address is not zero, or if the
binary is loaded in 16KiB compatibility mode.

Trying to read .note.gnu.property from the wrong address may cause a
segmentation fault while loading the binary, or it may result in
PROT_BTI not being applied to an otherwise BTI-compatible binary.

Bug: 424598884
Test: atest linker-unit-tests
Change-Id: Idaa3356f7368905987ced7f9fefb98a5f045e7bf
Signed-off-by: Csanád Hajdú <csanad.hajdu@arm.com>
[ Fix trivial merged conflicts - Kalesh Singh ]
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>