Refactor ScopedTrace.
The optimization I did broke the disable trace code since the call to
should_trace() happens even if tracing is disabled. So refactor so
that should_trace() is not called if tracing is disabled.

Also, small refactor so that the force_trace variable isn't needed
at the small cost of slightly duplicating code.

Test: Bionic unit test that crashes with tracing enabled in Scudo
Test: does not crash.
Change-Id: Ib9c29643d1870fc1d8486b51e934a3b5a8da743f