stdio: add the missing locking to _fwalk().
This ensures that fflush(nullptr) -- which flushes all open streams,
because the 1980s were a crazy time and no-one had threads yet -- holds
the lock while iterating the stream list.

Also rename the lock to make it clear that it's specifically the stream
list lock. (We might want to rename both of these at some point, but
probably after we get rid of the "glue" blocks entirely, or decide we're
not going to do that.)

Bug: http://b/130189834
Change-Id: Ieb349315831b99d3afbeae85da5402a25d6dd931