<string.h>: write our own strcspn()/strspn().
The massive speedup from using our optimized arm64 strchrnul() here
for the most common case suggests there's the potential for these to
be significantly faster for the other cases if we had some vector/simd
code here.

In the meantime, this is faster for all inputs than the nested loop
traditional BSD code (which does actually do pretty well for a single
delimiter, at the expense of poor performance with multiple delimiters)
and faster than the obvious bitset code. Though somewhere hboehm is
muttering about me trashing the data cache for my microbenchmark, I
think it's probably justified here because (a) it's the stack, (b) it's
only 256 bytes (and you're probably only hitting the first 64), and (c)
you're calling these functions because you're doing string processing,
and the input -- rather than the delimiters -- is where your data cache
is going.

Bug: http://b/437059814
Change-Id: Ifa441e2818422c32d671b5b15827748e8f876e79