Remove stdlib inlines.
Most of the similar functions already just use __RENAME() so that the compiler emits a call directly to the "real" function. Do that more consistently (also for the integer strto*() functions, not just the floating point ones). One less bit of cruft for readers to understand and fewer files for the preprocessor to wade through.

Change-Id: I5e1fbb573b046030867ed2d6788b9626521b50e7