libc: add portable-simd, start using strlen.cpp from it
This adds a foundation for us to build upon for portable simd
implementations of core functions. The goal is not for this code to
**match or exceed** the performance of hand-crafted assembly, but to
get us 80% of the way to hand-crafted asm with 20% of the coding and
maintenance effort.

That said, in the case of strlen, on a brya device (our primary x86_64
AL device) using Bionic's benchmarks:
- The SSE implementation of this essentially matches the hand-crafted
  asm in Bionic.
- The AVX2 implementation of this is _nearly_ a strict win (it's
  consistently <1ns slower on 32 byte strings, and approaches a 1.25x
  speedup as strings get longer).

Specific numbers are in
https://docs.google.com/document/d/1dT2hBKHP6zXPibAI_d77whqMVJ5WEFGGPRmAc3RRnEU/edit?tab=t.0 .

I've proven the concept of this _general_ simd pattern scaling to ~all
functions that we have optimized asm for, though the code is very ugly,
and probably subtly buggy (cl/815818901 - the tests/benchmarks were LLM
generated; very useful for catching quite a few bugs, but I make no
claims of perfection). I did this for SSE, AVX2, and AVX512 on my
personal machine.

I started with strlen because it's functionally quite simple, and there
was already a hand-crafted asm implementation in Bionic. The largest
impact of this (IMO) will be for the set of functions where we lack
that. That said, all of this can easily - I hope - be made to apply to
32-bit targets, and arches like ARM/RISCV.

Bug: 452714218
Test: Benchmarked on Brya

Change-Id: Ia301fb894ed869ff2fe478f3070b881800a3df21