More stdio benchmarks.
There were claims that our stdio performance depended on the fopen()
mode, so add the missing benchmarks to prove/disprove that.

From the benchmarks, it looks like there's some noise that can cause
misleading results for any of the options, but if you run repeatedly
they are indeed all pretty close in performance, as you'd expect.

Bug: http://b/124458286
Change-Id: Iaf33bf393e903fe4d26c5306bbc83b7c65809dc7