libc: add memrchr to portable-simd
memchr's implementation has many cool tricks, so refactor that to work
in terms of a 'Traits' object that encapsulates the diff in
functionality that we need for memrchr.

This is admittedly kinda footgun-y in a few ways (need to remember to
advance() a pointer rather than `ptr += x;`), but seems better to accept
some extra mental overhead rather than fully duplicating the 'driver'
logic with a few tweaks.

Bug: 453997143
Test: Benchmarks; go/bionic-psimd-benchmarks?tab=t.tcbvau7zpp9q

Change-Id: I46784dc246dc5a4c654c6b7ad71cec8c52d17931