arm64 setjmp: fix cookie check.
Also simplify the logic a bit by using some of the callee-save temporary
registers before we overwrite them, rather than using the stack, which
requires CFI directives and made this quite hard to read.

Unlike x86/x86-64, the checksum isn't the last word in the jmp_buf, it's
the last _used_ word in the jmp_buf, so the test needs to be updated
to work for arm64 too (but this version should work for any architecture).

Change-Id: I6fa3ed8c1d8e1ba61a323a51b6ac338ea5b0f373