call_once(): prepare to move to <stdlib.h>.
C23 moved this to <stdlib.h> (and C++26 is arranging to _not_ have it in <cstdlib>, for those who like to collect the differences between the .h> and <c headers). Before we do that, let's make it more usable (since it's now a lot less niche than the new <threads.h> header) and remove the need for a separate file of <stdlib.h> inlines and corresponding .cpp file.

(I did consider this at the time we added <threads.h>, but there are sufficiently few functions that could just be an ELF alias, and at the time no reason to assume <threads.h> would see any real traction given that <pthread.h> predates it by several decades, but <stdlib.h> is a lot more visible, and thus seems more likely to be used.)

Change-Id: I23809bc61127feb07474763596349b3e253c725e