Clear SME ZA state in clone() and vfork()
AAPCS64 defines SME ZA interface as private for clone(), fork() and
vfork() as well. This means that ZA state on entry can be "dormant" or
"off", while on return it can be unchanged or "off".
https://github.com/ARM-software/abi-aa/blob/main/aapcs64/aapcs64.rst#za-interfaces
Handling the "dormant" state would make the code unnecessary complex,
so for simplicity turn ZA state off on entry which ensures that the
state on return will be "off" as well.
It is not expected to be a common scenario anyway to call clone(),
fork() or vfork() in the middle of an SME operation.
Also, it would make very little sense for the "child" to start in
"dormant" state, as this state means SME is active but not in use with
lazy saving scheme in place, so it would expect somewhere to return and
restore ZA context.

This change also creates functional parity with glibc.

Test: Bionic unit tests pass, including all old and new SME related
tests.
Cherrypick-From: https://partner-android-review.googlesource.com/q/commit:c7adcbf1ecc0ea763bd20a1e041eafa8d2ed3789
Merged-In: Id5971e43c83f35d28fd3114d4e6bf1bcf9b638d5
Change-Id: Id5971e43c83f35d28fd3114d4e6bf1bcf9b638d5