fenv_test: be a bit clearer about FE_TONEARESTFROMZERO.
This is a bit confusing, because if you search other architectures' docs for the IEEE name (roundTiesToAway) for this rounding mode, you will see references to it. What you can't (currently) do on arm64 or x86-64 is set the floating point status register so that this is the default. Only riscv64 has that. (IEEE mandates it for _decimal_ floating point, because it's most useful for accounting, but it's optional for _binary_ floating point.)

Change-Id: I8540f3c840fb7ceb0bec4fd0cc6f2c469e843653